GetMaxListFor: use size_t and add a default like the other fields.

+++ modules/protocol/inspircd.cpp
@@ -114,7 +114,7 @@ public:
 		return maxhost ? maxhost : IRCDProto::GetMaxHost();
 	}
 
-	unsigned GetMaxListFor(Channel *c, ChannelMode *cm) override
+	size_t GetMaxListFor(Channel *c, ChannelMode *cm) override
 	{
 		ListLimits *limits = maxlist.Get(c);
 		if (limits)
